Add metadata for job queues
