+++ misk-admin/src/main/kotlin/misk/web/metadata/all/AllMetadataModule.kt
@@ -26,6 +26,7 @@ import misk.web.metadata.webaction.WebActionsMetadataProvider
  */
 class AllMetadataModule : KAbstractModule() {
   override fun configure() {
+    // Expose metadata as an API
     install(WebActionModule.create<AllMetadataAction>())
 
     // Built in metadata
@@ -34,7 +35,7 @@ class AllMetadataModule : KAbstractModule() {
     install(MetadataModule(JvmMetadataProvider()))
     install(MetadataModule(WebActionsMetadataProvider()))
 
-    // Install dashbaord tab
+    // Install dashboard tab
     install(WebActionModule.create<MetadataTabIndexAction>())
     install(DashboardModule.createHotwireTab<AdminDashboard, AdminDashboardAccess>(
       slug = "metadata",
