+++ misk-aws/src/test/kotlin/misk/jobqueue/sqs/SqsJobQueueTestModule.kt
@@ -36,8 +36,7 @@ class SqsJobQueueTestModule(
             queue_attribute_importer_frequency_ms = 0
           )
         )
-      )
-        .with(SqsTestModule(credentials, client))
+      ).with(SqsTestModule(credentials, client))
     )
   }
 }
