+++ misk-jobqueue/src/testFixtures/kotlin/misk/jobqueue/FakeJobQueueModule.kt
@@ -1,10 +1,13 @@
 package misk.jobqueue
 
 import misk.inject.KAbstractModule
+import misk.web.metadata.MetadataModule
 
 class FakeJobQueueModule : KAbstractModule() {
   override fun configure() {
     bind<JobQueue>().to<FakeJobQueue>()
     bind<TransactionalJobQueue>().to<FakeJobQueue>()
+
+    install(MetadataModule(JobqueueMetadataProvider()))
   }
 }
