+++ samples/exemplar/build.gradle.kts
@@ -38,6 +38,7 @@ dependencies {
   implementation(project(":misk-prometheus"))
   implementation(project(":misk-service"))
   implementation(project(":misk-testing"))
+  implementation(testFixtures(project(":misk-jobqueue")))
 
   testImplementation(libs.assertj)
   testImplementation(libs.awsDynamodb)
