+++ samples/exemplar/src/main/kotlin/com/squareup/exemplar/ExemplarService.kt
@@ -6,6 +6,7 @@ import misk.MiskRealServiceModule
 import misk.config.ConfigModule
 import misk.config.MiskConfig
 import misk.environment.DeploymentModule
+import misk.jobqueue.ExampleJobQueuesTestingModule
 import misk.metrics.backends.prometheus.PrometheusMetricsServiceModule
 import misk.monitoring.MonitoringModule
 import misk.web.MiskWebModule
@@ -23,6 +24,7 @@ fun main(args: Array<String>) {
     ExemplarMetadataModule(),
     ExemplarWebActionsModule(),
     ExemplarCronModule(),
+    ExampleJobQueuesTestingModule(),
     MiskRealServiceModule(),
     MiskWebModule(config.web),
     PrometheusMetricsServiceModule(config.prometheus),
